the routing structure is created (repeat!).
